requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
pydantic==2.12.5
requests==2.33.1
dnspython==2.8.0
python-whois==0.9.6
shodan==1.31.0
python-dotenv==1.2.2